refactor(cli): move onboarding openshell version helpers
